Added Raise To Top/Lower To Bottom menu items; implemented double-click to enter subcircuit; added support for alt to do interpolation when adding curve, made alt/shift work as expected when moving handle on an existing curve, rectangle, rounded rectangle, or oval; added more refined version for determining point containment for text; added control-arrow key configuration of Pin's Label Location attribute.
git-svn-id: https://circuit.svn.sourceforge.net/svnroot/circuit/trunk@75 70edf91d-0b9e-4248-82e7-2488d7716404

+++ src/com/cburch/draw/model/AbstractCanvasObject.java
@@ -195,7 +195,7 @@ public abstract class AbstractCanvasObject
 			if (color != null && color.getAlpha() == 0) {
 				return false;
 			} else {
-				GraphicsUtil.switchToWidth(g, width);
+				GraphicsUtil.switchToWidth(g, width.intValue());
 				if (color != null) g.setColor(color);
 				return true;
 			}
